Adhere to AbilityInputID changes by extending the base enum.
Update AbilitySystemSetup plugin with new ASSAbilityInputID. Have STAbilityInputID extend ASSAbilityInputID as required.

And reset the DefaultInput ini config settings and clean up the DefaultGameplayTags ini config settings.

+++ ShooterTest/Source/ShooterTest/Private/Character/Abilities/Interact/STGameplayAbility_CharacterInteract.cpp
@@ -12,7 +12,7 @@
 USTGameplayAbility_CharacterInteract::USTGameplayAbility_CharacterInteract(const FObjectInitializer& ObjectInitializer)
 	: Super(ObjectInitializer)
 {
-	AbilityInputID = ESTAbilityInputID::Interact;
+	AbilityInputID = static_cast<uint8>(ESTAbilityInputID::Interact);
 }
 
 void USTGameplayAbility_CharacterInteract::ASSOnAvatarSet(const FGameplayAbilityActorInfo* ActorInfo, const FGameplayAbilitySpec& Spec)
