+++ ShooterTest/Source/ShooterTest/Private/Character/Abilities/STGameplayAbility_CharacterJumpStatic.cpp
@@ -13,7 +13,7 @@
 USTGameplayAbility_CharacterJumpStatic::USTGameplayAbility_CharacterJumpStatic(const FObjectInitializer& ObjectInitializer)
 	: Super(ObjectInitializer)
 {
-	AbilityInputID = ESTAbilityInputID::Jump;
+	AbilityInputID = static_cast<uint8>(ESTAbilityInputID::Jump);
 	InstancingPolicy = EGameplayAbilityInstancingPolicy::NonInstanced;
 	bReplicateInputDirectly = true;		// bReplicateEndAbility in EndAbility() when replicating to the server doesn't always work because client's ability most likely isn't confirmed yet. So we do this bool instead to tell the server to run EndAbility(). (the better alternative to this bool is to use the input tasks but we can't because this is a non-instanced ability)
 	AbilityTags.AddTag(STNativeGameplayTags::Ability_Movement_JumpStatic);
